requirements.txt
transformers==4.38.2
torch==2.4.1
flask==3.0.3
flask_cors==5.0.0
torchvision==0.19.1
torchaudio==2.4.1
python-dotenv==1.0.1
